Refactor backup loader
